refactor: extract MenuBarView to separate file
- Extract MenuBarView from CallTranscriptionApp.swift to Views/MenuBarView.swift
- Improves code organization and testability
- Follows architectural pattern established by SettingsView
- Re-apply configuration fixes that were reset by xcodegen
  - Info.plist privacy descriptions
  - Entitlements for sandbox and audio access

Addresses code review feedback
